% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_enrich.R
\name{do_fisher_tests}
\alias{do_fisher_tests}
\title{Perform Fisher's tests}
\usage{
do_fisher_tests(gene.counts, bg.counts)
}
\arguments{
\item{gene.counts}{table; The term counts for the genes of interest.}

\item{bg.counts}{table; The term counts for the background.}
}
\value{
A vector with the p-values for each term.
}
\description{
Carry out the required over-representation tests for each of the terms annotating the genes of interest.
}
\examples{
# Generate a hypotetical scenario of annotations
bg.annotations <- sample(x = base::letters, size = 100, replace = TRUE)
gene.annotations <- sample(bg.annotations, size = 30)

# Compute the frequence of terms
bg.counts <- table(bg.annotations)
gene.counts <- table(gene.annotations)

#Perform Fisher's tests of over-representation for each term
pvals <- do_fisher_tests(gene.counts = gene.counts, bg.counts = bg.counts)

}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
