% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_enrich.R
\name{fun_enrich}
\alias{fun_enrich}
\title{Perform BP, CC, MF and REACTOME enrichment analyses}
\usage{
fun_enrich(gene.list, background, id.type = "ENTREZID", benjamini = F)
}
\arguments{
\item{gene.list}{character; A vector with the genes of interest.}

\item{background}{character; A vector with the background list of genes.}

\item{id.type}{character; One of ENTREZID (default), SYMBOL or UNIPROT accession. This is the ID type of \code{gene.list} and \code{background}.}

\item{benjamini}{logical; Whether to include Benjamini-Hochberg adjusted p-values or not.}
}
\value{
A list with four data frames, one per enrichment analysis:
\item{bp}{Contains the Biological Process \code{go.id}s annotating the genes of interest, together with their \code{term} 
description, p-values \code{pval} and Benjamini-Hochberg adjusted p-values \code{bh} if requested.}
\item{cc}{Same as \code{bp} but for Cellular Compartment.}
\item{mf}{Same as \code{bp} but for Molecular Function.}
\item{reactome}{Same as the rest but the first column is \code{react.id} instead of \code{go.id}.}
}
\description{
Given a list of genes of interest and a reference background, performs a Fisher's test for 
the over-representation of Gene Ontology terms and REACTOME pathways in the former.
}
\examples{
# Use the included lists of disease genes and genes associated with metabolic disorders
# as background and genes of interest, respectively
analysis <- fun_enrich(gene.list = metabolic, background = disease.genes, 
                      id.type = "ENTREZID", benjamini = TRUE)

}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
