% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_enrich.R
\name{plot_fun_enrich}
\alias{plot_fun_enrich}
\title{Plot the results of the BP, CC, MF and/or REACTOME enrichment analyses}
\usage{
plot_fun_enrich(enr, aspect = "ALL", benjamini = FALSE, top = 5,
  char_per_line = 80)
}
\arguments{
\item{enr}{listr; The list of data frames resulting from \code{fun_enrich}.}

\item{aspect}{character; One of ALL, BP, CC, MF or REACTOME.}

\item{benjamini}{logical; If TRUE, plot corrected p-values.}

\item{top}{integer; The number of top enriched terms to consider.}

\item{char_per_line}{integer; Controls the number of characters per line for
the term names in the plots.}
}
\value{
A ggplot object.
}
\description{
Given the output of \code{fun_enrich}, generates a bar plot of the -log10 of
the p-values for one or all analysed aspects.
}
\examples{
# Use the included lists of disease genes and genes associated with metabolic
# disorders as background and genes of interest, respectively
analysis <- fun_enrich(gene.list = metabolic, background = disease.genes, 
                      id.type = "ENTREZID", benjamini = TRUE)
# Plot the 
vis <- plot_fun_enrich(enr = analysis, aspect = "ALL", benjamini = TRUE, 
                       top = 5, char_per_line = 80)

}
\author{
Gregorio Alanis-Lobato \email{galanisl@uni-mainz.de}
}
